refactored teach layout user authentication check
