+++ frontend/app/routes/teach.js
@@ -1,5 +1,16 @@
 import Route from '@ember/routing/route';
-// import { inject as service } from '@ember/service';
-// import { tracked } from '@glimmer/tracking';
+import { inject as service } from '@ember/service';
 
-export default class TeachRoute extends Route {}
+export default class TeachRoute extends Route {
+  @service me;
+
+  beforeModel(transition) {
+    if (!this.me.isAuthenticated) {
+      // eslint-disable-next-line ember/no-controller-access-in-routes
+      let loginController = this.controllerFor('login');
+      loginController.set('previousTransition', transition);
+      this.transitionTo('login');
+    }
+    return super.beforeModel(transition);
+  }
+}
