+++ frontend/app/routes/teach/create.js
@@ -4,17 +4,6 @@ import { inject as service } from '@ember/service';
 export default class TeachCreateRoute extends Route {
   @service me;
 
-  beforeModel(transition) {
-    if (!this.me.isAuthenticated) {
-      if (!this.me.isAuthenticated) {
-        // eslint-disable-next-line ember/no-controller-access-in-routes
-        let loginController = this.controllerFor('login');
-        loginController.set('previousTransition', transition);
-      }
-      this.transitionTo('login');
-    }
-  }
-
   model() {
     return this.store.createRecord('chapter', {
       creator: this.me.get('user'),
